+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Removed
+
+
+## [0.18.1] 2020-12-01
+
+### Added
+
 * Added URDF and XML writers.
 * Added `compas.robots.RobotModel.to_urdf_file`.
 * Added `compas.files.URDF.from_robot`.